refactor: extract reusable resource resolver
