+++ src/Resolver/ArticleTeaserResolver.php
@@ -6,16 +6,8 @@ namespace Atoolo\GraphQL\Search\Resolver;
 
 use Atoolo\GraphQL\Search\Types\ArticleTeaser;
 use Atoolo\GraphQL\Search\Types\Asset;
-use Atoolo\GraphQL\Search\Types\Image;
-use Atoolo\GraphQL\Search\Types\ImageCharacteristic;
-use Atoolo\GraphQL\Search\Types\ImageSource;
-use Atoolo\Resource\Resource;
-use Atoolo\Resource\ResourceHierarchyLoader;
-use Atoolo\Resource\ResourceHierarchyWalker;
 use DateTime;
-use InvalidArgumentException;
 use Overblog\GraphQLBundle\Definition\ArgumentInterface;
-use Psr\Log\LoggerInterface;
 
 /**
  * @phpstan-type ImageData array{
@@ -38,167 +30,28 @@ use Psr\Log\LoggerInterface;
 class ArticleTeaserResolver implements Resolver
 {
     public function __construct(
-        private readonly UrlRewriter $urlRewriter,
-        private readonly LoggerInterface $logger,
-        private readonly ResourceHierarchyLoader $hierarchyLoader
+        private readonly ResourceAssetResolver $assetResolver,
+        private readonly ResourceKickerResolver $kickerResolver,
+        private readonly ResourceDateResolver $dateResolver
     ) {
     }
 
     public function getKicker(
         ArticleTeaser $teaser
     ): ?string {
-        return $this->getKickerFromResource($teaser->resource);
+        return $this->kickerResolver->getKicker($teaser->resource);
     }
-
-    public function getKickerFromResource(
-        Resource $resource,
-    ): ?string {
-        $kickerText = $resource->data->getString(
-            'base.teaser.kicker',
-            $resource->data->getString('base.kicker')
-        );
-        if (!empty($kickerText)) {
-            return $kickerText;
-        }
-        $kickerText = null;
-        $walker = new ResourceHierarchyWalker($this->hierarchyLoader);
-        $walker->init($resource);
-        while ($parent = $walker->primaryParent()) {
-            $kickerText = $parent->data->getString('base.kicker');
-            if (!empty($kickerText)) {
-                break;
-            }
-        }
-        return $kickerText;
-    }
-
 
     public function getDate(
         ArticleTeaser $teaser
     ): ?DateTime {
-        return $this->getDateFromResource($teaser->resource);
-    }
-
-    public function getDateFromResource(
-        Resource $resource,
-    ): ?DateTime {
-        $timestamp = $resource->data->getInt('base.teaser.date');
-        $date = new DateTime();
-        $date->setTimestamp($timestamp);
-        return $date;
+        return $this->dateResolver->getDate($teaser->resource);
     }
 
     public function getAsset(
         ArticleTeaser $teaser,
         ArgumentInterface $args
     ): ?Asset {
-        return $this->getAssetFromResource($teaser->resource, $args);
-    }
-
-    public function getAssetFromResource(
-        Resource $resource,
-        ArgumentInterface $args
-    ): ?Asset {
-
-        /** @var ImageData|array{} $imageData */
-        $imageData = $resource->data->getAssociativeArray(
-            'base.teaser.image'
-        );
-        if (empty($imageData)) {
-            return null;
-        }
-
-        try {
-            $characteristic = ImageCharacteristic::valueOfCamelCase(
-                $imageData['characteristic'] ?? 'normal'
-            );
-        } catch (InvalidArgumentException $e) {
-            $this->logger->error(
-                'Invalid characteristic for teaser image',
-                [
-                    'resource' => $resource->location,
-                    'exception' => $e
-                ]
-            );
-            return null;
-        }
-
-        $copyright = $imageData['copyright'] ?? null;
-        $alternativeText = $imageData['text'] ?? null;
-        $caption = $imageData['legend'] ?? null;
-        $description = $imageData['description'] ?? null;
-        $original = null;
-        if (isset($imageData['original'])) {
-            $original = $this->toImageSource(
-                'original',
-                $imageData['original']
-            );
-        }
-
-        /** @var string $variant */
-        $variant = $args['variant'];
-
-        $sources = [];
-        if (
-            isset($imageData['variants']) &&
-            is_array($imageData['variants'][$variant])
-        ) {
-            $sources = $this->toImageSourceList(
-                $variant,
-                $imageData['variants'][$variant]
-            );
-        }
-
-        return new Image(
-            $copyright,
-            $caption,
-            $description,
-            $alternativeText,
-            $original,
-            $characteristic,
-            $sources
-        );
-    }
-
-    /**
-     * @param array<ImageSourceData> $variantData
-     * @return ImageSource[]
-     */
-    private function toImageSourceList(
-        string $variant,
-        array $variantData
-    ): array {
-        $sources = [];
-        foreach ($variantData as $sourceData) {
-            $sources[] = $this->toImageSource(
-                $variant,
-                $sourceData
-            );
-        }
-        return $sources;
-    }
-
-    /**
-     * @param ImageSourceData $sourceData
-     */
-    private function toImageSource(
-        string $variant,
-        array $sourceData
-    ): ImageSource {
-        $url = $this->urlRewriter->rewrite(
-            UrlRewriterType::IMAGE,
-            $sourceData['url']
-        );
-        $width = $sourceData['width'];
-        $height = $sourceData['height'];
-        $mediaQuery = $sourceData['mediaQuery'] ?? null;
-
-        return new ImageSource(
-            $variant,
-            $url,
-            $width,
-            $height,
-            $mediaQuery
-        );
+        return $this->assetResolver->getAsset($teaser->resource, $args);
     }
 }
