+++ src/Resolver/NewsTeaserResolver.php
@@ -12,25 +12,21 @@ use Overblog\GraphQLBundle\Definition\ArgumentInterface;
 class NewsTeaserResolver implements Resolver
 {
     public function __construct(
-        private readonly ArticleTeaserResolver $articleTeaserResolver
+        private readonly ResourceAssetResolver $assetResolver,
+        private readonly ResourceDateResolver $dateResolver
     ) {
     }
 
     public function getDate(
         NewsTeaser $teaser
     ): ?DateTime {
-        return $this->articleTeaserResolver->getDateFromResource(
-            $teaser->resource
-        );
+        return $this->dateResolver->getDate($teaser->resource);
     }
 
     public function getAsset(
         NewsTeaser $teaser,
         ArgumentInterface $args
     ): ?Asset {
-        return $this->articleTeaserResolver->getAssetFromResource(
-            $teaser->resource,
-            $args
-        );
+        return $this->assetResolver->getAsset($teaser->resource, $args);
     }
 }
